requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-extensions
uritemplate
pyyaml